+++ demo/src/main/java/ebookstore/com/demo/storage/StorageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StorageRepository extends JpaRepository<Storage, Long> {
+public interface StorageRepository extends JpaRepository<Storage, Integer> {
 
     List<Storage> findByName(String name);
 }
